generation/requirements.txt
numpy==1.26.4
rich==13.4.2
tqdm==4.66.5
scipy==1.11.4
scikit-learn==1.3.0
opencv-python==4.8.0.76
imageio==2.21.1
imageio-ffmpeg==0.4.8
omegaconf==2.3.0
einops==0.4.1
torch==2.1.1
huggingface_hub==0.19.4
diffusers==0.24.0
accelerate==0.24.1
transformers==4.38.0
bitsandbytes==0.43.0
rembg[gpu,cli]==2.0.49
xformers==0.0.23
gradio==4.19.2
open-clip-torch==2.24.0
plyfile==1.0.3
loguru==0.7.2

# FastAPI for generation service
fastapi==0.115.6
uvicorn==0.34.0

# CLIP validator (for competitive mining)
git+https://github.com/openai/CLIP.git

# Additional PIL support
Pillow==10.1.0